# Doc/src/main/java/json/Api-Docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Online_Shop_API",
    "description": "This is a Website API programmed by Java Spring Boot by Fattah Arbab",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/panel/slider/swap-up/{id}": {
      "put": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "moveUp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/slider/swap-down/{id}": {
      "put": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "moveDown",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/nav/swap-up/{id}": {
      "put": {
        "tags": [
          "nav-panel-controller"
        ],
        "operationId": "moveUp_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/nav/swap-down/{id}": {
      "put": {
        "tags": [
          "nav-panel-controller"
        ],
        "operationId": "moveDown_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/slider": {
      "get": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPanelResponseListSliderDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/nav": {
      "get": {
        "tags": [
          "nav-panel-controller"
        ],
        "operationId": "getAll_5",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPanelResponseListNavBarDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "boolean"
          }
        }
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SliderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "orderNumber": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          }
        }
      },
...
      "NavBarDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "orderNumber": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiPanelResponseListSliderDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SliderDto"
            }
          },
          "totalCounts": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiPanelResponseListNavBarDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NavBarDto"
            }
          },
          "totalCounts": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}